refactoring server;working prod deployment
