+++ common/database/kustomize/overlays/prod/.env
@@ -1,0 +1,1 @@
+POSTGRES_DB=ccg-production-db
