--- common/database/kustomize/overlays/prod/config.properties
@@ -1,2 +1,0 @@
-POSTGRES_DB="open-ims-db"
-POSTGRES_PASSWORD="password"
