+++ common/database/kustomize/overlays/prod/postgres-credentials-patch.json
@@ -1,0 +1,12 @@
+[
+    {
+        "op": "replace",
+        "path": "/spec/template/spec/containers/0/env/0/valueFrom/secretKeyRef/name",
+        "value": "prod-postgres-credentials"
+    },
+    {
+        "op": "replace",
+        "path": "/spec/template/spec/containers/0/env/1/valueFrom/secretKeyRef/name",
+        "value": "prod-postgres-credentials"
+    }
+]
