+++ common/database/kustomize/overlays/prod/postgres-credentials.yaml
@@ -1,0 +1,7 @@
+# apiVersion: v1
+# kind: Secret
+# metadata:
+#   name: postgres-credentials
+# stringData:
+#   POSTGRES_USER: 
+#   POSTGRES_PASSWORD:
