--- common/database/kustomize/overlays/staging/config.properties
@@ -1,2 +1,0 @@
-POSTGRES_DB="open-ims-db"
-POSTGRES_PASSWORD="UNSERCUREPASS"
