--- common/database/kustomize/overlays/staging/kustomization.yaml
@@ -1,21 +1,0 @@
-resources:
-- ../../base
-
-namePrefix: staging-
-
-configMapGenerator:
-  - name: postgres-config
-    env: config.properties
-
-patches:
-- path: init-container-patch.json
-  target:
-    group: apps
-    version: v1
-    kind: StatefulSet
-    name: postgres-stateful-set
-
-images:
-- name: postgres
-  newName: postgres
-  newTag: "15.3"
