+++ common/ingress/overlays/prod/env-name-patch.json
@@ -2,11 +2,11 @@
     {
         "op": "replace",
         "path": "/spec/rules/0/http/paths/0/backend/service/name",
-        "value": "staging-client-cluster-ip-service"
+        "value": "prod-client-cluster-ip-service"
     },
     {
         "op": "replace",
         "path": "/spec/rules/0/http/paths/1/backend/service/name",
-        "value": "staging-server-cluster-ip-service"
+        "value": "prod-server-cluster-ip-service"
     }
 ]
