+++ common/ingress/overlays/prod/kustomization.yaml
@@ -2,3 +2,11 @@ resources:
 - ../../base
 
 namePrefix: prod-
+
+patches:
+- path: env-name-patch.json
+  target:
+    group: networking.k8s.io
+    version: v1
+    kind: Ingress
+    name: ingress-service
