--- common/ingress/overlays/staging/kustomization.yaml
@@ -1,12 +1,0 @@
-resources:
-- ../../base
-
-namePrefix: staging-
-
-patches:
-- path: env-name-patch.json
-  target:
-    group: networking.k8s.io
-    version: v1
-    kind: Ingress
-    name: ingress-service
