+++ microservices/client/kustomize/overlays/prod/kustomization.yaml
@@ -1,9 +1,13 @@
 resources:
   - ../../base
 
-# namePrefix: 
-#   dev-
+namePrefix: 
+  prod-
 
 configMapGenerator:
-  - name: open-ims-client-map
+  - name: client-config
     env: config.properties
+
+images:
+- name: fcabrera01/open-ims-client
+  newName: fcabrera01/ccg-client
