--- microservices/client/kustomize/overlays/staging/config.properties
@@ -1,1 +1,0 @@
-CLIENT_ENV_VAR="Woo! ENV vars in STAGING!"
