--- microservices/client/kustomize/overlays/staging/kustomization.yaml
@@ -1,13 +1,0 @@
-resources:
-  - ../../base
-
-namePrefix: 
-  staging-
-
-configMapGenerator:
-  - name: client-config
-    env: config.properties
-
-images:
-- name: fcabrera01/open-ims-client
-  newName: fcabrera01/open-ims-client-staging
