+++ microservices/server/kustomize/overlays/prod/.env
@@ -1,0 +1,4 @@
+PGDATABASE=ccg-production-db
+PGPORT=5432
+PGHOST=prod-postgres-stateful-set-0.prod-postgres-headless-service
+SERVER_PORT=3000
