--- microservices/server/kustomize/overlays/prod/config.properties
@@ -1,1 +1,0 @@
-SERVER_ENV_VAR="Woo! ENV vars in PROD!"
