--- microservices/server/kustomize/overlays/staging/config.properties
@@ -1,4 +1,0 @@
-PGUSER='postgres'
-PGHOST='postgres-headless-service'
-PGPORT='5432'
-PGDATABASE='open-ims-db'
