--- microservices/server/kustomize/overlays/staging/kustomization.yaml
@@ -1,13 +1,0 @@
-resources:
-  - ../../base
-
-namePrefix: 
-  staging-
-
-configMapGenerator:
-  - name: server-config
-    env: config.properties
-
-images:
-- name: fcabrera01/open-ims-server
-  newName: fcabrera01/open-ims-server-staging
