+++ microservices/server/src/controllers/test.ts
@@ -1,9 +1,5 @@
 import { Express, Request, Response } from "express"
-import { query } from "../db/queries"
-import { clientCheck } from "../db/helpers";
-
-// Any complex business logic, please make a function that handles this. We don't want to overbloat our controllers.
-// This can include external api calls, complex queries, and data manipulation.
+import { query } from "../db"
 
 
 export const getClicks = async (req: Request, res: Response) => {
@@ -13,6 +9,19 @@ export const getClicks = async (req: Request, res: Response) => {
     maxNum = result?.rows[0].maxval;
 
     res.json({ maxNum: maxNum });
+}
+
+export const seedDB = async (req: Request, res: Response) => {
+    try {
+        await query('CREATE TABLE IF NOT EXISTS exampletable (nums integer UNIQUE)');
+        await query('INSERT INTO exampletable(nums) VALUES($1)', [1]);
+        console.log("Your DB has been seeded!");
+        res.json({ msg: "Successfully seeded the Database with the test data!" });
+    } catch (error) {
+        console.log("There was an error seeding the DB! (Most likely a dup key error, because the DB has already been seeded!) see below:");
+        console.log(error);
+        res.json({ msg: error });
+    }
 }
 
 export const addClick = async (req: Request, res: Response) => {
@@ -25,17 +34,4 @@ export const addClick = async (req: Request, res: Response) => {
     await query('INSERT INTO exampletable(nums) VALUES($1)', [maxNum + 1]);
 
     res.json({ success: true });
-}
-
-export const checkPoolClient = (req: Request, res: Response) => {
-    try {
-        console.log("Printing the amount of idle clients below (Pool of 10 total):");
-        console.log("--------------");
-        console.log(clientCheck());
-        console.log("--------------");
-        res.json({ success: true });
-    } catch (error) {
-        console.log(`Error trying to fetch clientCheck from DB!`);
-        res.json({ success: false });
-    }
 }
