+++ microservices/server/src/routes/test.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import { 
     getClicks, 
     addClick,
-    checkPoolClient 
+    seedDB
 } from '../controllers/test';
 import catchAsync from '../utils/catchAsync';
 
@@ -15,8 +15,8 @@ router.route('/api/server/getClickedNum')
 router.route('/api/server/addClick')
     .get(catchAsync(addClick))
 
-router.route('/api/server/poolclientcheck')
-    .get(checkPoolClient)
+router.route('/api/server/seed')
+    .get(catchAsync(seedDB))
 
 
 export default router;
